pyclass: tidy up frozen implementation
