+++ src/impl_/pycell.rs
@@ -1,0 +1,2 @@
+//! Externally-accessible implementation of pycell
+pub use crate::pycell::impl_::{GetBorrowChecker, PyClassMutability};
