+++ src/impl_/pyclass.rs
@@ -2,7 +2,8 @@ use crate::{
     exceptions::{PyAttributeError, PyNotImplementedError},
     ffi,
     impl_::freelist::FreeList,
-    pycell::{GetBorrowChecker, PyCellLayout, PyClassMutability},
+    impl_::pycell::{GetBorrowChecker, PyClassMutability},
+    pycell::PyCellLayout,
     pyclass_init::PyObjectInit,
     type_object::PyLayout,
     Py, PyAny, PyCell, PyClass, PyErr, PyMethodDefType, PyNativeType, PyResult, PyTypeInfo, Python,
