+++ tests/ui/abi3_nativetype_inheritance.stderr
@@ -1,29 +1,4 @@
 error[E0277]: the trait bound `PyDict: PyClass` is not satisfied
- --> tests/ui/abi3_nativetype_inheritance.rs:5:1
-  |
-5 | #[pyclass(extends=PyDict)]
-  | ^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `PyClass` is not implemented for `PyDict`
-  |
-  = help: the trait `PyClass` is implemented for `TestClass`
-  = note: required because of the requirements on the impl of `PyClassBaseType` for `PyDict`
-  = note: this error originates in the attribute macro `pyclass` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-error[E0277]: the trait bound `PyDict: PyClass` is not satisfied
-   --> tests/ui/abi3_nativetype_inheritance.rs:5:1
-    |
-5   | #[pyclass(extends=PyDict)]
-    | ^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `PyClass` is not implemented for `PyDict`
-    |
-    = help: the trait `PyClass` is implemented for `TestClass`
-    = note: required because of the requirements on the impl of `PyClassBaseType` for `PyDict`
-note: required by a bound in `PyRefMut`
-   --> src/pycell.rs
-    |
-    | pub struct PyRefMut<'p, T: PyClass<Frozen = False>> {
-    |                                    ^^^^^^^^^^^^^^ required by this bound in `PyRefMut`
-    = note: this error originates in the attribute macro `pyclass` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-error[E0277]: the trait bound `PyDict: PyClass` is not satisfied
     --> tests/ui/abi3_nativetype_inheritance.rs:5:1
      |
 5    | #[pyclass(extends=PyDict)]
@@ -37,3 +12,13 @@ note: required by a bound in `ThreadCheckerInherited`
      | pub struct ThreadCheckerInherited<T: PyClass + Send, U: PyClassBaseType>(
      |                                                         ^^^^^^^^^^^^^^^ required by this bound in `ThreadCheckerInherited`
      = note: this error originates in the attribute macro `pyclass` (in Nightly builds, run with -Z macro-backtrace for more info)
+
+error[E0277]: the trait bound `PyDict: PyClass` is not satisfied
+ --> tests/ui/abi3_nativetype_inheritance.rs:5:1
+  |
+5 | #[pyclass(extends=PyDict)]
+  | ^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `PyClass` is not implemented for `PyDict`
+  |
+  = help: the trait `PyClass` is implemented for `TestClass`
+  = note: required because of the requirements on the impl of `PyClassBaseType` for `PyDict`
+  = note: this error originates in the attribute macro `pyclass` (in Nightly builds, run with -Z macro-backtrace for more info)
